# Conflicts:
#	Libs/fmod/include/fmod.h
#	Projects/UnitTests/Classes/Tests/CPUCountTest.cpp
#	Sources/CMake/ConfigureFiles/DavaConfigTemplate.in
#	Sources/CMake/Modules/CMake-common.cmake
#	Sources/CMake/Modules/FindQT5.cmake
#	Sources/CMake/Modules/PlatformSettings.cmake
#	Sources/Internal/CMakeLists.txt
#	Sources/Internal/Platform/TemplateAndroid/ExternC/AndroidLayer.cpp
#	Sources/Internal/Platform/TemplateWin32/DeviceInfoWinUAP.h
#	Sources/Internal/Platform/TemplateWin32/WinUAPXamlApp.cpp
#	Sources/Internal/Render/2D/TextBlockDistanceRender.cpp
#	Sources/Internal/Render/Image/LibDdsHelper.cpp
#	Sources/Internal/Render/RenderState.cpp
#	Sources/Internal/Scene3D/Entity.cpp
#	Sources/Internal/Scene3D/Entity.h
#	Sources/Internal/UI/Layouts/FlowLayoutAlgorithm.cpp
#	Sources/Internal/UI/UIControlSystem.cpp
#	Sources/Internal/UI/UIScrollView.cpp
#	Sources/Internal/UI/UIScrollViewContainer.cpp
#	Tools/QuickEd/Classes/GameCore.cpp
#	Tools/ResourceEditor/Classes/Qt/Main/mainwindow.cpp
#	Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureBrowser.cpp
#	Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureListDelegate.cpp


Former-commit-id: af658b11a6660e9b2d653d19782a81917c3545ce

+++ Tools/QuickEd/Classes/Model/ControlProperties/ValueProperty.cpp
@@ -53,14 +53,14 @@ ValueProperty::~ValueProperty()
 
 }
 
-int ValueProperty::GetCount() const
+uint32 ValueProperty::GetCount() const
 {
-    return static_cast<int>(children.size());
+    return static_cast<int32>(children.size());
 }
 
-AbstractProperty *ValueProperty::GetProperty(int index) const
+AbstractProperty* ValueProperty::GetProperty(int32 index) const
 {
-    if (0 <= index && index < children.size())
+    if (0 <= index && index < static_cast<int32>(children.size()))
     {
         return children[index];
     }
